backend/requirements.txt
pyarrow==11.0.0
datasets==2.16.1
weaviate-client==3.26.2
python-dotenv==1.0.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
numpy==1.26.4
pandas==2.2.2
gunicorn

